+++ functions/src/index.ts
@@ -1,8 +1,7 @@
 import { initializeApp } from 'firebase-admin/app';
 
-// Firebase Admin SDK 초기화
 initializeApp();
 
 export * from './github';
 export * from './schedule';
-export * from './hyperclovax';
+export * from './clova';
